backend/pyproject.toml
[project]
name = "anvil-backend"
version = "0.1.0"
description = "Anvil — NVMe Validator & IOps Lab, FastAPI backend"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Anvil contributors" }]

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "pydantic-settings>=2.5",
  "sqlalchemy[asyncio]>=2.0",
  "alembic>=1.14",
  "asyncpg>=0.30",
  "aiosqlite>=0.20",
  "orjson>=3.10",
  "structlog>=24.4",
  "httpx>=0.27",
  "jinja2>=3.1",
  "python-ulid>=3.0",
  "websockets>=13.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "ruff>=0.7",
  "mypy>=1.13",
  "types-setuptools",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["anvil*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"anvil/api/*" = ["B008"]
"alembic/versions/*" = ["E402"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
